Gene and Chromosome classes refactoring

+++ src/test/java/genetic/util/ConverterTest.java
@@ -76,15 +76,15 @@ public class ConverterTest {
         List<Note> noteList = Arrays.asList(notes);
 
         Chromosome chromosome = Converter.fromNotes(noteList, 16);
-        assertThat(chromosome.getGene(0).getValue(), is((short) 33));
-        assertThat(chromosome.getGene(1).getValue(), is((short) Constants.REST.value()));
-        assertThat(chromosome.getGene(9).getValue(), is((short) 98));
-        assertThat(chromosome.getGene(13).getValue(), is((short) Constants.REST.value()));
+        assertThat(chromosome.getGene(0).getValue(), is(33));
+        assertThat(chromosome.getGene(1).getValue(), is(Constants.REST.value()));
+        assertThat(chromosome.getGene(9).getValue(), is(98));
+        assertThat(chromosome.getGene(13).getValue(), is(Constants.REST.value()));
         for (int i = 2; i < 16; i++) {
             if (i == 9 || i == 13) {
                 continue;
             }
-            assertThat(chromosome.getGene(i).getValue(), is((short) Constants.TENUTO.value()));
+            assertThat(chromosome.getGene(i).getValue(), is(Constants.TENUTO.value()));
         }
     }
